Move AST rotation implementation to own file
